Move and split field
