+++ django_bitcoin/fields/utils.py
@@ -1,51 +1,5 @@
-#
-# Django field type for a Bitcoin Address
-#
-import re
-from django import forms
-from django.forms.util import ValidationError
 import hashlib
-
-
-class BCAddressField(forms.CharField):
-    default_error_messages = {
-        'invalid': 'Invalid Bitcoin address.',
-        }
-
-    def __init__(self, *args, **kwargs):
-        super(BCAddressField, self).__init__(*args, **kwargs)
-
-    def clean(self, value):
-        if not value and not self.required:
-            return None
-
-        if not value.startswith(u"1") and not value.startswith(u"3"):
-            raise ValidationError(self.error_messages['invalid'])
-        value = value.strip()
-
-        if "\n" in value:
-            raise ValidationError(u"Multiple lines in the bitcoin address")
-
-        if " " in value:
-            raise ValidationError(u"Spaces in the bitcoin address")
-
-        if re.match(r"[a-zA-Z1-9]{27,35}$", value) is None:
-            raise ValidationError(self.error_messages['invalid'])
-        version = get_bcaddress_version(value)
-        if version is None:
-            raise ValidationError(self.error_messages['invalid'])
-        return value
-
-
-def is_valid_btc_address(value):
-    value = value.strip()
-    if re.match(r"[a-zA-Z1-9]{27,35}$", value) is None:
-        return False
-    version = get_bcaddress_version(value)
-    if version is None:
-        return False
-    return True
-
+import re
 
 __b58chars = '123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz'
 __b58base = len(__b58chars)
@@ -57,7 +11,7 @@ def b58encode(v):
 
     long_value = 0L
     for (i, c) in enumerate(v[::-1]):
-        long_value += (256**i) * ord(c)
+        long_value += (256 ** i) * ord(c)
 
     result = ''
     while long_value >= __b58base:
@@ -70,17 +24,20 @@ def b58encode(v):
     # leading 0-bytes in the input become leading-1s
     nPad = 0
     for c in v:
-        if c == '\0': nPad += 1
-        else: break
+        if c == '\0':
+            nPad += 1
+        else:
+            break
 
-    return (__b58chars[0]*nPad) + result
+    return (__b58chars[0] * nPad) + result
+
 
 def b58decode(v, length):
     """ decode v into a string of len bytes
     """
     long_value = 0L
     for (i, c) in enumerate(v[::-1]):
-        long_value += __b58chars.find(c) * (__b58base**i)
+        long_value += __b58chars.find(c) * (__b58base ** i)
 
     result = ''
     while long_value >= 256:
@@ -91,10 +48,12 @@ def b58decode(v, length):
 
     nPad = 0
     for c in v:
-        if c == __b58chars[0]: nPad += 1
-        else: break
+        if c == __b58chars[0]:
+            nPad += 1
+        else:
+            break
 
-    result = chr(0)*nPad + result
+    result = chr(0) * nPad + result
     if length is not None and len(result) != length:
         return None
 
@@ -106,7 +65,7 @@ def b36encode(number, alphabet='0123456789abcdefghijklmnopqrstuvwxyz'):
     if not isinstance(number, (int, long)):
         long_value = 0L
         for (i, c) in enumerate(number[::-1]):
-            long_value += (256**i) * ord(c)
+            long_value += (256 ** i) * ord(c)
         number = long_value
 
     base36 = '' if number != 0 else '0'
@@ -126,14 +85,26 @@ def b36decode(number):
     return int(number, 36)
 
 
+def is_valid_btc_address(value):
+    value = value.strip()
+    if re.match(r"[a-zA-Z1-9]{27,35}$", value) is None:
+        return False
+    version = get_bcaddress_version(value)
+    if version is None:
+        return False
+    return True
+
+
 def get_bcaddress_version(strAddress):
     """ Returns None if strAddress is invalid.    Otherwise returns integer version of address. """
     addr = b58decode(strAddress, 25)
-    if addr is None: return None
+    if addr is None:
+        return None
     version = addr[0]
     checksum = addr[-4:]
-    vh160 = addr[:-4] # Version plus hash160 is what is checksummed
-    h3=hashlib.sha256(hashlib.sha256(vh160).digest()).digest()
+    vh160 = addr[:-4]  # Version plus hash160 is what is checksummed
+    h3 = hashlib.sha256(hashlib.sha256(vh160).digest()).digest()
     if h3[0:4] == checksum:
         return ord(version)
     return None
+
